Refactor code and update game logic: Fix typo in ft_free.c and ft_check_map.c
This commit fixes a typo in the ft_free.c and ft_check_map.c files. In ft_free.c, the variable name "hight" has been corrected to "height" in the while loop condition. In ft_check_map.c, the variable name "hight" has been corrected to "height" in the while loop condition and in the ft_check_outline function. These changes ensure that the correct variable is used and improve the accuracy of the code.

Note: The recent user commits and repository commits have been considered to generate this commit message.